% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_forest_profile.R
\name{build_forest_profile}
\alias{build_forest_profile}
\title{Build a 2D Forest Profile of L-System Trees Based on Inventory Data}
\usage{
build_forest_profile(
  data,
  n_trees,
  height,
  diameter,
  label,
  string = NULL,
  angle = 15,
  crown_diameter = NULL,
  h_reduction = (1 + sqrt(5))/2 - 1,
  d_reduction = (1 + sqrt(5))/2 - 1,
  randomness = FALSE,
  angle_cv = 0.1,
  length_cv = 0.1,
  leaf_size = NULL,
  n_iter = 6,
  dist = 3,
  dist_cv = 0.3
)
}
\arguments{
\item{data}{A data frame that represents field or simulated forest inventory data.}

\item{n_trees}{Integer. The number of sample trees.}

\item{height}{Name of the column containing the heights of the trees.}

\item{diameter}{Name of the column containing the dbh or base diameters of the trees.}

\item{label}{A column with the labels that identifies the trees (scientific names or other taxonomic identification).}

\item{string}{The string or the column of the strings that contains the turtle graphics instructions, created or not by \code{iterate_lsystem} function. If NULL (default), the binary tree model is used.}

\item{angle}{Numeric. The angle in degrees when a change in direction is requested.}

\item{crown_diameter}{Crown diameter of the plant.}

\item{h_reduction}{Numeric. Reduction rate to be applied to the length of the branches whenever the symbols '(' and ')' appear in the string. Default is the golden ratio.}

\item{d_reduction}{Numeric. Reduction rate to be applied to the diameter of the branches whenever the symbols '(' and ')' appear in the string. Default is the golden ratio.}

\item{randomness}{Logical. Should randomness be applied to the angle of the branches?}

\item{angle_cv}{Numerical. The coefficient of variation of the angle, used when \code{randomness} = TRUE. Default is 0.1.}

\item{length_cv}{Numerical. The coefficient of variation of the length, used when \code{randomness} = TRUE. Default is 0.1.}

\item{leaf_size}{Size of leaves (branches from the ends of the plant). Default is NULL (follows the applied d_reduction value).}

\item{n_iter}{Integer. The number of iterations to build the binary tree model when \code{string} = NULL (default). Default is 6.}

\item{dist}{Numerical. The average linear distance between the trees.}

\item{dist_cv}{Numerical. The coefficient of variation of the \code{dist}argument. Default is .3.}
}
\description{
Build a 2D Forest Profile of L-System Trees Based on Inventory Data
}
\details{
A column of \code{data} can be declared in the string argument. In this case the column must contain for each line (tree) the turtle graphics string that builds the tree.
}
