% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_tree}
\alias{build_tree}
\title{Build a 2D L-System Tree Data Frame Using Turtle Graphics}
\usage{
build_tree(
  string = NULL,
  height = NULL,
  diameter = NULL,
  h_reduction = (1 + sqrt(5))/2 - 1,
  d_reduction = (1 + sqrt(5))/2 - 1,
  angle = 90,
  randomness = FALSE,
  angle_cv = 0.1,
  length_cv = 0.1,
  leaf_size = NULL
)
}
\arguments{
\item{string}{A character vector giving the strings containing the turtle graphics instructions. Created by iterate_lsystem function.}

\item{height}{Total height of the plant.}

\item{diameter}{Base diameter of the plant.}

\item{h_reduction}{Numeric. Reduction rate to be applied to the length of the branches whenever the symbols '(' and ')' appear in the string. Default is the golden ratio.}

\item{d_reduction}{Numeric. Reduction rate to be applied to the diameter of the branches whenever the symbols '(' and ')' appear in the string. Default is the golden ratio.}

\item{angle}{Numeric. The angle in degrees when a change in direction is requested}

\item{randomness}{Logical. Should randomness be applied to the angle of the branches?}

\item{angle_cv}{Numerical. The coefficient of variation of the angle, used when randomness = TRUE. Default is 0.1,}

\item{length_cv}{Numerical. The coefficient of variation of the length, used when randomness = TRUE. Default is 0.1,}

\item{leaf_size}{Size of leaves (branches from the ends of the plant). Default is NULL (follows the applied d_reduction value).}
}
\description{
Build a 2D L-System Tree Data Frame Using Turtle Graphics
}
