% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tree.R
\name{draw_tree}
\alias{draw_tree}
\title{Visualize a fractal tree}
\usage{
draw_tree(
  tree_df,
  d_col = NULL,
  d_factor = 1,
  branch_color = "black",
  leaf_color = "black"
)
}
\arguments{
\item{tree_df}{The resulting data.frame from the \code{build_tree} function.}

\item{d_col}{Name of the column that contains the diameters along the branches. Usually 'diameter'.}

\item{d_factor}{Numeric. Multiplication factor to harmonize the visualization of the diameters.}

\item{branch_color}{Character. Color of the tree branches}

\item{leaf_color}{Character. Color of the tree leaves.}
}
\description{
Visualize a fractal tree
}
