% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractalforest-package.R
\docType{package}
\name{fractalforest-package}
\alias{fractalforest}
\alias{fractalforest-package}
\title{fractalforest: Simulate Fractal Trees and Forests}
\description{
Create and visualize fractal trees and fractal forests, based on the Lindenmayer system (L-system). For more details see Lindenmayer (1968a) \doi{10.1016/0022-5193(68)90079-9} and Lindenmayer (1968b) \doi{10.1016/0022-5193(68)90080-5}.
}
\author{
\strong{Maintainer}: Sergio Costa \email{sergio.vscf@gmail.com} (\href{https://orcid.org/0000-0001-5432-317X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Gilmara Machado \email{gmachado@unicentro.br} (\href{https://orcid.org/0000-0003-1455-9086}{ORCID}) [copyright holder]
}

}
\keyword{internal}
