% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_point_cloud.R
\name{generate_point_cloud}
\alias{generate_point_cloud}
\title{Generate a point cloud from 2D trees data frame}
\usage{
generate_point_cloud(tree_df, pt_distance = 0.3)
}
\arguments{
\item{tree_df}{The resulting data.frame from the \code{build_tree} or \code{build_forest_profile} functions.}

\item{pt_distance}{Numeric. Distance between points in the point cloud.}
}
\description{
Generate a point cloud from 2D trees data frame
}
