% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_matrix.R
\name{get_tree_matrix}
\alias{get_tree_matrix}
\title{Extract the fractal tree matrix from a fractal tree data frame}
\usage{
get_tree_matrix(tree_df)
}
\arguments{
\item{tree_df}{The resulting data.frame from the \code{build_tree} function.}
}
\description{
Extract the fractal tree matrix from a fractal tree data frame
}
