% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_branch_symbol.R
\name{is_branch_symbol}
\alias{is_branch_symbol}
\title{Check if character is a branch symbol}
\usage{
is_branch_symbol(x)
}
\arguments{
\item{x}{an R object to be tested.}
}
\description{
Check if character is a branch symbol
}
