% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_lsystem.R
\name{iterate_lsystem}
\alias{iterate_lsystem}
\title{Rewrite an Axiom Using Production Rules to Give a String Ready for Turtle Graphics}
\usage{
iterate_lsystem(init = NULL, rules = NULL, n = 5, verbose = 0L)
}
\arguments{
\item{init}{A character string giving variables (symbols) to use as the initial string, also known as the axiom.}

\item{rules}{A data frame containing columns "inp" and "out". These contain the input variables and the corresponding replacement string.}

\item{n}{An integer giving the number of cycles or iterations desired.}

\item{verbose}{An integer giving the level of information desired as the calculation proceeds. verbose = 1L gives basic information at each cycle. Any value greater than 1 gives much more detail. Supress messages by using a value less than 1.}
}
\description{
Rewrite an Axiom Using Production Rules to Give a String Ready for Turtle Graphics
}
