% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest_profile.R
\name{plot_forest_profile}
\alias{plot_forest_profile}
\title{Visualize the 2D profile of a fractal trees forest}
\usage{
plot_forest_profile(
  tree_df,
  d_col = NULL,
  branch_color = "black",
  leaf_color = "black",
  simplify = F,
  dTolerance = 0.15
)
}
\arguments{
\item{tree_df}{The resulting data.frame from the \code{build_forest_profile} function.}

\item{d_col}{Name of the column that contains the diameters along the branches. Usually 'diameter'.}

\item{branch_color}{Character. Color of the tree branches.}

\item{leaf_color}{Character. Color of the tree leaves.}

\item{simplify}{Logical. Whether to simplify geometries of the tree. Default to FALSE.}

\item{dTolerance}{Numeric. If simplify = TRUE, the tolerance parameter to be considered. Default to .15.}
}
\description{
Visualize the 2D profile of a fractal trees forest
}
