% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_matrix.R
\name{get_tree_matrix}
\alias{get_tree_matrix}
\title{Extract the Fractal Tree Matrix from a Tree Data Frame}
\usage{
get_tree_matrix(tree_df)
}
\arguments{
\item{tree_df}{A data frame containing the tree data, typically resulting from the \code{build_tree} function. The data frame should include columns \code{from_x}, \code{from_y}, \code{to_x}, and \code{to_y}, which represent the coordinates of the tree segments.}
}
\value{
A matrix of unique coordinates (x, y) representing the positions of the tree's segments.
}
\description{
Extract the Fractal Tree Matrix from a Tree Data Frame
}
\details{
This function extracts the start and end coordinates of the tree segments and returns the result as a matrix.
}
