% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_branch_symbol.R
\name{is_branch_symbol}
\alias{is_branch_symbol}
\title{Check if character is a branch symbol}
\usage{
is_branch_symbol(x)
}
\arguments{
\item{x}{an R object to be tested.}
}
\value{
A logical vector the same length as \code{x},
with \code{TRUE} for elements that are \emph{not} one of the branch symbols
(\code{"["}, \code{"]"}, \code{"+"}, \code{"-"}, \code{"("}, \code{")"}, \code{"<"}, \code{">"}),
and \code{FALSE} otherwise.
}
\description{
Check if character is a branch symbol
}
