% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pop_symbol.R
\name{is_pop_symbol}
\alias{is_pop_symbol}
\title{Check if character is a pop symbol}
\usage{
is_pop_symbol(x)
}
\arguments{
\item{x}{an R object to be tested.}
}
\value{
A logical vector the same length as \code{x},
with \code{TRUE} for elements equal to the pop symbol \code{"]"},
and \code{FALSE} otherwise.
}
\description{
Check if character is a pop symbol
}
